+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 dev
 ---
 
-No changes.
+* Added a new parameter: -g/--user-agent. It is used to specify the user-agent
+  string that is sent by the scripts.
 
 0.1 (2015-02-09)
 ----------------